Almost done refactor.  Just need to separate state from service.
